Move TraceXXX (not TracingState) into new recorder.py
