+++ guppylang-internals/src/guppylang_internals/definition/traced.py
@@ -43,7 +43,7 @@ from guppylang_internals.error import GuppyComptimeError
 from guppylang_internals.metadata.common import FunctionMetadata, add_metadata
 from guppylang_internals.nodes import GlobalCall
 from guppylang_internals.span import SourceMap
-from guppylang_internals.tracing.state import (
+from guppylang_internals.tracing.recorder import (
     TraceCall,
     TraceFunctionLoad,
     TraceLoad,
@@ -59,7 +59,7 @@ from guppylang_internals.tys.ty import InputFlags, Type, UnitaryFlags, type_to_r
 PyFunc = Callable[..., Any]
 
 if TYPE_CHECKING:
-    from guppylang_internals.tracing.state import Trace
+    from guppylang_internals.tracing.recorder import Trace
 
 
 @dataclass(frozen=True)
