+++ guppylang-internals/src/guppylang_internals/tracing/object.py
@@ -25,11 +25,8 @@ from guppylang_internals.engine import DEF_STORE, ENGINE
 from guppylang_internals.error import GuppyComptimeError, GuppyError, GuppyTypeError
 from guppylang_internals.frame_util import get_calling_frame
 from guppylang_internals.ipython_inspect import normalize_ipython_dummy_files
-from guppylang_internals.tracing.state import (
-    TraceWire,
-    get_tracing_state,
-    tracing_active,
-)
+from guppylang_internals.tracing.recorder import TraceWire
+from guppylang_internals.tracing.state import get_tracing_state, tracing_active
 from guppylang_internals.tracing.util import capture_guppy_errors, hide_trace
 from guppylang_internals.tys.ty import EnumType, FunctionType, StructType, Type
